Further backend typescript integration, need clarification on some data structures to continue
