+++ server/controller/RouteType.ts
@@ -1,0 +1,7 @@
+import { Request, Response, NextFunction } from 'express';
+
+export type RouteType = (
+    req: Request,
+    res: Response,
+    next: NextFunction,
+ ) => void
