+++ server/routes/recipeRoute.ts
@@ -5,12 +5,12 @@ const databaseController = require('../controller/databaseController');
 const dalleImageController = require('../controller/dalleImageController');
 
 // Scrape info for a specific URL provided
-router.get('/scrapeurl', scrapingController, (req, res, next) => {
+router.get('/scrapeurl', scrapingController, ( req: Request, res: Response, next : NextFunction)=> {
   res.status(200).json(res.locals);
 });
 
 // Get all recipes stored in the database
-router.get('/all', databaseController.getAllRecipes, (req, res, next) => {
+router.get('/all', databaseController.getAllRecipes, ( req: Request, res: Response, next : NextFunction) => {
   res.status(200).json(res.locals);
 });
 
@@ -19,13 +19,13 @@ router.post(
   '/add',
   dalleImageController.generateImage,
   databaseController.addRecipe,
-  (req, res, next) => {
+  ( req: Request, res: Response, next : NextFunction)=> {
     res.status(200).json(res.locals);
   }
 );
 
 // Update a recipe (except image) with a specific id in the database. The request shall have the recipe id in url parameter and all other information (except imagePath) in the request body.
-router.put('/update/:id', databaseController.updateRecipe, (req, res, next) => {
+router.put('/update/:id', databaseController.updateRecipe, ( req: Request, res: Response, next : NextFunction) => {
   res.status(200).json(res.locals);
 });
 
@@ -34,7 +34,7 @@ router.put(
   '/update/image/:id',
   dalleImageController.generateImage,
   databaseController.updateImage,
-  (req, res, next) => {
+  ( req: Request, res: Response, next : NextFunction) => {
     res.status(200).json(res.locals);
   }
 );
@@ -43,12 +43,12 @@ router.put(
 router.delete(
   '/delete/:id',
   databaseController.deleteRecipe,
-  (req, res, next) => {
+  ( req: Request, res: Response, next : NextFunction) => {
     res.status(200).json('The recipe has been successfully deleted.');
   }
 );
 
-router.get('/userrecipe/:id', databaseController.getUserRecipe, (req, res) => {
+router.get('/userrecipe/:id', databaseController.getUserRecipe, ( req: Request, res: Response) => {
   res.status(200).json(res.locals);
 })
